requirements.txt
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.9
langchain-openai==0.1.9
langchain_text_splitters==0.2.1
python-dotenv==1.0.1
pymupdf==1.24.5
python-socketio==5.11.2
openai==1.35.0
qdrant-client==1.9.2
pydantic==1.10.9
fastapi==0.111.0
reportlab==4.2.2
uvicorn[standard]
